@@ -18,14 +18,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
   get: ^4.1.4
-  hexcolor: ^2.0.3
+  hexcolor: ^2.0.4
 
 dev_dependencies:
   flutter_test:
@@ -47,6 +47,7 @@ flutter:
     - assets/images/
     - assets/images/icons/
     - assets/fonts/
+    - assets/texts/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
