@@ -1,69 +1,63 @@
-import 'dart:ui';
-
-import 'package:compod_app/commons/app_colors.dart';
-import 'package:compod_app/commons/app_fonts.dart';
+import 'package:compod_app/commons/compod_images.dart';
+import 'package:compod_app/home/controller/home_controller.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class HomeTile extends StatefulWidget {
-  HomeTile(this.image, this.text, {Key key}) : super(key: key);
+  HomeTile({required this.type, required this.text, Key? key}) : super(key: key);
 
-  final String image;
+  final HomeTileType type;
   final String text;
+  static const size = 160.0;
+  static const imageSize = 64.0;
 
   @override
   _HomeTileState createState() => _HomeTileState();
 }
 
 class _HomeTileState extends State<HomeTile> {
-  final double _size = 160;
-  final double _imageSize = 64;
-
-  bool _highlighted = false;
-
-  void _buttonPressed() {}
+  var _highlighted = false;
 
   @override
   Widget build(BuildContext context) {
+    final HomeController controller = Get.find();
+
     return Container(
       decoration: BoxDecoration(
         boxShadow: [BoxShadow(color: Colors.black45, blurRadius: 2, spreadRadius: 2, offset: Offset(0, 2))],
-        border: Border.all(color: _highlighted ? AppColors.green.withOpacity(0.6) : Colors.white.withOpacity(0.6), width: 2),
+        border: Border.all(color: _highlighted ? Get.theme.colorScheme.secondaryVariant.withOpacity(0.6) : Colors.white.withOpacity(0.6), width: 2),
         borderRadius: BorderRadius.circular(5),
-        gradient: LinearGradient(begin: Alignment(0, 0.4), end: Alignment.bottomCenter, colors: [AppColors.buttonBlue1, AppColors.buttonBlue2]),
-        // color: AppColors.buttonBlue,
+        gradient: LinearGradient(begin: Alignment(0, 0.4), end: Alignment.bottomCenter, colors: [Get.theme.primaryColor, Get.theme.primaryColorDark]),
       ),
-      width: _size,
-      height: _size,
+      width: HomeTile.size,
+      height: HomeTile.size,
       child: MaterialButton(
-        onPressed: _buttonPressed,
-        onHighlightChanged: (highlight) {
-          setState(() {
-            _highlighted = highlight;
-          });
-        },
+        onPressed: () => controller.tileAction(widget.type),
+        onHighlightChanged: (highlight) => setState(() => _highlighted = highlight),
         child: Column(
           children: [
             SizedBox(height: 32),
-            Image.asset(widget.image, width: _imageSize, height: _imageSize),
-            // Icon(Icons., size: _imageSize),
-            Expanded(
-              child: Center(
-                child: Text(
-                  widget.text,
-                  textAlign: TextAlign.center,
-                  style: TextStyle(
-                      height: 1,
-                      decoration: TextDecoration.none,
-                      fontSize: 18,
-                      color: Colors.white,
-                      fontFamily: AppFonts.PalanquinDark,
-                      fontWeight: FontWeight.w600),
-                ),
-              ),
-            )
+            Image.asset(widget.type.image, width: HomeTile.imageSize, height: HomeTile.imageSize),
+            Expanded(child: Center(child: Text(widget.text.toUpperCase(), textAlign: TextAlign.center, style: Get.textTheme.button?.copyWith(height: 1.5))))
           ],
         ),
       ),
     );
+  }
+}
+
+enum HomeTileType {
+  Doctor,
+  Heart,
+  Donate,
+}
+
+extension on HomeTileType {
+  String get image {
+    switch (this) {
+      case HomeTileType.Doctor: return AppImages.Doctor;
+      case HomeTileType.Donate: return AppImages.Donate;
+      case HomeTileType.Heart: return AppImages.Heart;
+    }
   }
 }
