@@ -1,8 +1,8 @@
-class AppFonts {
+class CompodFonts {
 
   static const PalanquinDark        = 'Palanquin-Dark';
   static const Palanquin            = 'Palanquin';
 
-  AppFonts._();
+  CompodFonts._();
 
 }
