@@ -3,7 +3,7 @@
 // // To perform an interaction with a widget in your test, use the WidgetTester
 // // utility that Flutter provides. For example, you can send tap and scroll
 // // gestures. You can also use WidgetTester to find child widgets in the widget
-// // tree, read text, and verify that the values of widget properties are correct.
+// // tree, read texts, and verify that the values of widget properties are correct.
 //
 // import 'package:flutter/material.dart';
 // import 'package:flutter_test/flutter_test.dart';
@@ -16,15 +16,15 @@
 //     await tester.pumpWidget(MyApp());
 //
 //     // Verify that our counter starts at 0.
-//     expect(find.text('0'), findsOneWidget);
-//     expect(find.text('1'), findsNothing);
+//     expect(find.texts('0'), findsOneWidget);
+//     expect(find.texts('1'), findsNothing);
 //
 //     // Tap the '+' icon and trigger a frame.
 //     await tester.tap(find.byIcon(Icons.add));
 //     await tester.pump();
 //
 //     // Verify that our counter has incremented.
-//     expect(find.text('0'), findsNothing);
-//     expect(find.text('1'), findsOneWidget);
+//     expect(find.texts('0'), findsNothing);
+//     expect(find.texts('1'), findsOneWidget);
 //   });
 // }
